Restructure CSS ClassName To More Understable
