+++ src/App.js
@@ -1,16 +1,16 @@
 import { Routes, Route } from "react-router-dom";
 import Mockman from "mockman-js";
-import { Home } from "./pages/Home/Home";
-import { MenuItemList } from "./pages/MenuItemList/MenuItemList";
+import { HomePage } from "./pages/HomePage/HomePage";
+import { MenuItemListPage } from "./pages/MenuItemListPage/MenuItemListPage";
 
 function App() {
   return (
     <div className="App">
       <Routes>
-        <Route path="/" element={<Home />}></Route>
+        <Route path="/" element={<HomePage />}></Route>
         <Route
           path="/main_category/:mainCategoryId"
-          element={<MenuItemList />}
+          element={<MenuItemListPage />}
         ></Route>
         <Route path="/mockman" element={<Mockman />}></Route>
       </Routes>
