+++ src/backend/db/menuItems.js
@@ -415,7 +415,8 @@ export const menu_items = [
     sub_category_id: "SC_5",
     benefits: [],
     veg_egg_non: "veg",
-    image: "",
+    image:
+      "https://res.cloudinary.com/dcu6sympq/image/upload/v1683924536/grocery/cookies/17_hdvfuw.webp",
     rating: 4.0,
     delivery_time_in_mins: 20,
     item_variant: [
