+++ src/component/Banner/constant.js
