+++ src/pages/HomePage/Component/MainCategoryList/Constants.js
