+++ src/pages/HomePage/HomePage.css
@@ -1,3 +1,3 @@
-.HomeContainer {
+.HomePageContainer {
   margin: 0px 110px;
 }
