--- src/pages/MenuItemList/Component/SubCategoryWithProduct/SubCategoryWithProduct.jsx
@@ -1,15 +1,0 @@
-import "./SubCategoryWithProduct.css";
-
-import { ItemList } from "./Component/ItemList/ItemList";
-import { SubCategoryStack } from "./Component/SubCategoryStack/SubCategoryStack";
-import { FilterStack } from "./Component/FilterStack/FilterStack";
-
-export function SubCategoryWithProduct() {
-  return (
-    <div className="SubCategoryMenuItemContainer">
-      <SubCategoryStack />
-      <ItemList />
-      <FilterStack />
-    </div>
-  );
-}
