--- src/pages/MenuItemList/MenuItemList.css
@@ -1,3 +1,0 @@
-.MenuItemListContainer {
-  margin: 0px 110px;
-}
