+++ src/pages/MenuItemListPage/Component/MenuItemListWithFilters/MenuItemListWithFilters.css
@@ -1,4 +1,4 @@
-.SubCategoryMenuItemContainer {
+.MenuItemListWithFiltersContainer {
   margin-top: 136px;
   display: grid;
   grid-template-columns: 200px 1fr 200px;
