+++ src/pages/MenuItemListPage/Component/MenuItemListWithFilters/MenuItemListWithFilters.jsx
@@ -1,0 +1,15 @@
+import "./MenuItemListWithFilters.css";
+
+import { MenuItemList } from "./Component/MenuItemList/MenuItemList";
+import { SubCategoryStack } from "./Component/SubCategoryStack/SubCategoryStack";
+import { FilterStack } from "./Component/FilterStack/FilterStack";
+
+export function MenuItemListWithFilters() {
+  return (
+    <div className="MenuItemListWithFiltersContainer">
+      <SubCategoryStack />
+      <MenuItemList />
+      <FilterStack />
+    </div>
+  );
+}
