+++ src/pages/MenuItemListPage/MenuItemListPage.css
@@ -1,0 +1,3 @@
+.MenuItemListPageContainer {
+  margin: 0px 110px;
+}
